commons #101 refactor field value extraction
